[refactor] component presenter/container 분리
- component presenter/container 분리 시도
- 관심사의 분리가 선명하게 되지 못함
- 다른 방식의 디자인패턴 혹은 설계패턴 필요
