+++ src/components/Map/MapContainer.tsx
@@ -2,6 +2,7 @@ import React, { useState, useRef, useEffect } from 'react';
 import dotenv from 'dotenv';
 import mapboxgl from 'mapbox-gl';
 import 'mapbox-gl/dist/mapbox-gl.css';
+import UpperButtonsContainer from './ButtonGroup/UpperButtonsContainer';
 
 import MapPresenter from './MapPresenter';
 
@@ -66,13 +67,12 @@ function MapContainer(): React.ReactElement {
   };
 
   return (
-    <MapPresenter
-      fullscreenHandler={fullscreenHandler}
-      smallscreenHandler={smallscreenHandler}
-      mapRef={mapRef}
-      plusZoom={plusZoom}
-      minusZoom={minusZoom}
-    />
+    <MapPresenter mapRef={mapRef} plusZoom={plusZoom} minusZoom={minusZoom}>
+      <UpperButtonsContainer
+        fullscreenHandler={fullscreenHandler}
+        smallscreenHandler={smallscreenHandler}
+      />
+    </MapPresenter>
   );
 }
