+++ src/components/Sidebar/SidebarContentFewer/SidebarContentContainer.tsx
@@ -1,8 +1,13 @@
 import React from 'react';
 import SidebarContentPresenter from './SidebarContentPresenter';
+import SidebarContentThemeContainer from './SidebarContentThemeContainer';
 
 function SidebarContentContainer(): React.ReactElement {
-  return <SidebarContentPresenter />;
+  return (
+    <SidebarContentPresenter>
+      <SidebarContentThemeContainer />
+    </SidebarContentPresenter>
+  );
 }
 
 export default SidebarContentContainer;
