+++ src/components/Sidebar/SidebarContentMore/SidebarContentContainer.tsx
@@ -1,24 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import SidebarContentPresenter from './SidebarContentPresenter';
-import FeatureTypePresenter from './FeatureTypePresenter';
+import FeatureTypeContainer from './FeatureTypeContainer';
 
 function SidebarContentContainer(): React.ReactElement {
-  const [featureName, setFeatureName] = useState('');
-  const [styledFeatureList, setStyledFeatureList] = useState([]); // 추후 전역 상태로 변경
-
-  const featureClickHandler = (name: string) => {
-    if (name !== featureName) {
-      setFeatureName(name);
-    }
-  };
-
   return (
     <SidebarContentPresenter>
-      <FeatureTypePresenter
-        featureName={featureName}
-        styledFeatureList={styledFeatureList}
-        featureClickHandler={featureClickHandler}
-      />
+      <FeatureTypeContainer />
     </SidebarContentPresenter>
   );
 }
